refactor to remove V1
- updated documentation
- removed old files
also, the only reason i named it "V2" in the past was because i needed a clear differentiator between the old and the new, and it was a quick and concise way to do it.
however, in the future, if i create a new version of MenuHandler, i may reuse the v2 naming scheme just for the sake of simplicity. i could also use v3, but it would seem pretty weird for a jump in version to occur when the current doesn't have the version marked.